e2e: address retransmit-only settlement review feedback
- ClosestRetransmitOnlyDevice returns the flagged retransmit-only metro
  set so the test can skip only when no metro is flagged and fail (naming
  the metros) when metros are flagged but no reachable device matched,
  instead of a silent skip masking an unreachable feature under test.
- assert_subscribed_groups logs the seat's last-seen subscription state
  (retransmit/leader booleans + subscribed group pubkeys) on timeout so
  on-call can distinguish an oracle bug from a leader-exclusion bug.
- Extract the shared pay/open-phase/balance/withdraw settlement flow into
  runShredSettlement (qa_shred_settlement_test.go); TestQA_MulticastSettlement
  and TestQA_RetransmitOnlySettlement are now thin wrappers, so the
  self-heal/epoch-tail/escrow-drain/withdraw-retry machinery lives once.
